Finally, finally, finally fixing everything. Downgrading Udev to work with Amazon’s old–ass kernels.
See the comments in Nucleus/bundle.sh for more info, but it boils down to the
fact that Amazon runs kernels older than God, and that the latest Udev (145-1
and later, made it into the Arch [core] repository on 2008-8-26) won’t work
with any kernel earlier than 2.6.25. So, we grab the last working Udev from
the Arch Rollback Machine (you, my friend, are a very useful tool!) and work
from there.

It’s not really an ideal, nor a long–term, solution to this problem…
really, it’d be better if Amazon provided non–Jurassic kernels, or, better yet,
allowed us to provide our own. While dealing with all this today, I posted a
rant on this topic on the EC2 forums; let’s hope it gets some attention:
http://developer.amazonwebservices.com/connect/thread.jspa?threadID=36675